Split PSF photometry classes into new modules
